+++ src/core/methods/trackTransactions/helpers/checkTransactionStatus/getPendingTransactions.ts
@@ -1,26 +1,16 @@
 import { getIsTransactionPending } from 'store/actions/transactions/transactionStateByStatus';
 import { SignedTransactionType } from 'types/transactions.types';
 
-export interface PendingTransactionType {
-  hash: string;
-  previousStatus: string;
-}
-
 export function getPendingTransactions(
-  transactions: SignedTransactionType[],
-  timedOutHashes: string[]
-): PendingTransactionType[] {
+  transactions: SignedTransactionType[]
+): SignedTransactionType[] {
   const pendingTransactions = transactions.reduce(
-    (acc: PendingTransactionType[], { status, hash }) => {
+    (acc: SignedTransactionType[], transaction) => {
       if (
-        hash != null &&
-        !timedOutHashes.includes(hash) &&
-        getIsTransactionPending(status)
+        transaction.hash != null &&
+        getIsTransactionPending(transaction.status)
       ) {
-        acc.push({
-          hash,
-          previousStatus: status
-        });
+        acc.push(transaction);
       }
       return acc;
     },