+++ src/store/slices/network/networkSlice.types.ts
@@ -1,5 +1,5 @@
-import { CurrentNetworkType } from 'types/network.types';
+import { NetworkType } from 'types/network.types';
 
 export interface NetworkSliceType {
-  network: CurrentNetworkType;
+  network: NetworkType;
 }